refactor: split catalog helpers
